slovak translation update for upcomming version
* translation around 94%
* added `_()` call for some forgotten strings
* added `.` at the end of translation string for pyboard mode (all the
  other strings cotain the `.` at the end of the string)
* cast type to int in `editor.py`, because mu was unable to run on my
  system (fedora 35, python 3.10.1)

+++ mu/interface/editor.py
@@ -304,8 +304,8 @@ class EditorPane(QsciScintilla):
         # Make the margin left of line numbers follow zoom level
         self.setMarginWidth(0, margins[size])
         # Make margins around debugger-marker follow zoom level
-        self.setMarginWidth(1, margins[size] * 0.25)
-        self.setMarginWidth(4, margins[size] * 0.1)
+        self.setMarginWidth(1, int(margins[size] * 0.25))
+        self.setMarginWidth(4, int(margins[size] * 0.1))
 
     @property
     def label(self):
